FlaskApi/requirements.txt
Click==7.0
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.14
flask-swagger-ui==3.20.9
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
mysqlclient==1.4.6
PyMySQL==0.9.3
python-dotenv==0.10.5
PyYAML==5.3
six==1.14.0
SQLAlchemy==1.3.12
Werkzeug==0.16.0
